+++ src/k_rahCycle.cu
@@ -1,20 +1,20 @@
 #include "kernels.cuh"
 
-__device__ float a_d;
-__device__ float b_d;
+__device__ double a_d;
+__device__ double b_d;
 
-__global__ void d0_kernel(double *pai_d, double *d0_d, float CD1, float HGHT)
+__global__ void d0_kernel(double *pai_d, double *d0_d, double CD1, double HGHT)
 {
     unsigned int pos = threadIdx.x + blockIdx.x * blockDim.x;
 
     if (pos < width_d * height_d) {
-        float cd1_pai_root = sqrtf(CD1 * pai_d[pos]);
+        double cd1_pai_root = sqrtf(CD1 * pai_d[pos]);
 
         d0_d[pos] = HGHT * ((1.0 - (1.0 / cd1_pai_root)) + (pow(exp(1.0), -cd1_pai_root) / cd1_pai_root));
     }
 }
 
-__global__ void ustar_kernel_STEEP(double *zom_d, double *d0_d, double *ustar_d, float u10)
+__global__ void ustar_kernel_STEEP(double *zom_d, double *d0_d, double *ustar_d, double u10)
 {
     unsigned int pos = threadIdx.x + blockIdx.x * blockDim.x;
 
@@ -23,7 +23,7 @@ __global__ void ustar_kernel_STEEP(double *zom_d, double *d0_d, double *ustar_d,
     }
 }
 
-__global__ void ustar_kernel_ASEBAL(double *zom_d, double *ustar_d, float u200)
+__global__ void ustar_kernel_ASEBAL(double *zom_d, double *ustar_d, double u200)
 {
     unsigned int pos = threadIdx.x + blockIdx.x * blockDim.x;
 
@@ -32,17 +32,17 @@ __global__ void ustar_kernel_ASEBAL(double *zom_d, double *ustar_d, float u200)
     }
 }
 
-__global__ void zom_kernel_STEEP(double *d0_d, double *pai_d, double *zom_d, float A_ZOM, float B_ZOM)
+__global__ void zom_kernel_STEEP(double *d0_d, double *pai_d, double *zom_d, double A_ZOM, double B_ZOM)
 {
     unsigned int pos = threadIdx.x + blockIdx.x * blockDim.x;
 
-    float HGHT = 4;
-    float CD = 0.01;
-    float CR = 0.35;
-    float PSICORR = 0.2;
+    double HGHT = 4;
+    double CD = 0.01;
+    double CR = 0.35;
+    double PSICORR = 0.2;
 
     if (pos < width_d * height_d) {
-        float gama = pow((CD + CR * (pai_d[pos] / 2)), -0.5);
+        double gama = pow((CD + CR * (pai_d[pos] / 2)), -0.5);
         if (gama < 3.3)
             gama = 3.3;
 
@@ -50,7 +50,7 @@ __global__ void zom_kernel_STEEP(double *d0_d, double *pai_d, double *zom_d, flo
     }
 }
 
-__global__ void zom_kernel_ASEBAL(double *ndvi_d, double *albedo_d, double *zom_d, float A_ZOM, float B_ZOM)
+__global__ void zom_kernel_ASEBAL(double *ndvi_d, double *albedo_d, double *zom_d, double A_ZOM, double B_ZOM)
 {
     unsigned int pos = threadIdx.x + blockIdx.x * blockDim.x;
 
@@ -59,41 +59,41 @@ __global__ void zom_kernel_ASEBAL(double *ndvi_d, double *albedo_d, double *zom_
     }
 }
 
-__global__ void kb_kernel(double *zom_d, double *ustar_d, double *pai_d, double *kb1_d, double *ndvi_d, float ndvi_max, float ndvi_min)
+__global__ void kb_kernel(double *zom_d, double *ustar_d, double *pai_d, double *kb1_d, double *ndvi_d, double ndvi_max, double ndvi_min)
 {
     unsigned int pos = threadIdx.x + blockIdx.x * blockDim.x;
 
-    float HGHT = 4;
+    double HGHT = 4;
 
-    float VON_KARMAN = 0.41;
-    float visc = 0.00001461;
-    float pr = 0.71;
-    float c1 = 0.320;
-    float c2 = 0.264;
-    float c3 = 15.1;
-    float cd = 0.2;
-    float ct = 0.01;
-    float sf_c = 0.3;
-    float sf_d = 2.5;
-    float sf_e = 4.0;
-    float soil_moisture_day_rel = 0.33;
+    double VON_KARMAN = 0.41;
+    double visc = 0.00001461;
+    double pr = 0.71;
+    double c1 = 0.320;
+    double c2 = 0.264;
+    double c3 = 15.1;
+    double cd = 0.2;
+    double ct = 0.01;
+    double sf_c = 0.3;
+    double sf_d = 2.5;
+    double sf_e = 4.0;
+    double soil_moisture_day_rel = 0.33;
 
     if (pos < width_d * height_d) {
-        float fc = 1 - pow((ndvi_d[pos] - ndvi_max) / (ndvi_min - ndvi_max), 0.4631);
-        float fs = 1 - fc;
+        double fc = 1 - pow((ndvi_d[pos] - ndvi_max) / (ndvi_min - ndvi_max), 0.4631);
+        double fs = 1 - fc;
 
-        float Re = (ustar_d[pos] * 0.009) / visc;
-        float Ct = pow(pr, -0.667) * pow(Re, -0.5);
-        float ratio = c1 - c2 * (exp(cd * -c3 * pai_d[pos]));
-        float nec = (cd * pai_d[pos]) / (ratio * ratio * 2);
-        float kbs = 2.46 * pow(Re, 0.25) - 2;
+        double Re = (ustar_d[pos] * 0.009) / visc;
+        double Ct = pow(pr, -0.667) * pow(Re, -0.5);
+        double ratio = c1 - c2 * (exp(cd * -c3 * pai_d[pos]));
+        double nec = (cd * pai_d[pos]) / (ratio * ratio * 2);
+        double kbs = 2.46 * pow(Re, 0.25) - 2;
 
-        float kb1_fst_part = (cd * VON_KARMAN) / (4 * ct * ratio * (1 - exp(nec * -0.5)));
-        float kb1_sec_part = pow(fc, 2) + (VON_KARMAN * ratio * (zom_d[pos] / HGHT) / Ct);
-        float kb1_trd_part = pow(fc, 2) * pow(fs, 2) + kbs * pow(fs, 2);
-        float kb_ini = kb1_fst_part * kb1_sec_part * kb1_trd_part;
+        double kb1_fst_part = (cd * VON_KARMAN) / (4 * ct * ratio * (1 - exp(nec * -0.5)));
+        double kb1_sec_part = pow(fc, 2) + (VON_KARMAN * ratio * (zom_d[pos] / HGHT) / Ct);
+        double kb1_trd_part = pow(fc, 2) * pow(fs, 2) + kbs * pow(fs, 2);
+        double kb_ini = kb1_fst_part * kb1_sec_part * kb1_trd_part;
 
-        float SF = sf_c + (1 / (1 + exp(sf_d - sf_e * soil_moisture_day_rel)));
+        double SF = sf_c + (1 / (1 + exp(sf_d - sf_e * soil_moisture_day_rel)));
 
         kb1_d[pos] = kb_ini * SF;
     }
@@ -104,9 +104,9 @@ __global__ void aerodynamic_resistance_kernel_STEEP(double *zom_d, double *d0_d,
     unsigned int pos = threadIdx.x + blockIdx.x * blockDim.x;
 
     if (pos < width_d * height_d) {
-        float rah_fst_part = 1 / (ustar_d[pos] * VON_KARMAN);
-        float rah_sec_part = logf((10 - d0_d[pos]) / zom_d[pos]);
-        float rah_trd_part = rah_fst_part * kb1_d[pos];
+        double rah_fst_part = 1 / (ustar_d[pos] * VON_KARMAN);
+        double rah_sec_part = logf((10 - d0_d[pos]) / zom_d[pos]);
+        double rah_trd_part = rah_fst_part * kb1_d[pos];
         rah_d[pos] = (rah_fst_part * rah_sec_part) + rah_trd_part;
     }
 }
@@ -120,7 +120,7 @@ __global__ void aerodynamic_resistance_kernel_ASEBAL(double *ustar_d, double *ra
     }
 }
 
-__global__ void rah_correction_cycle_STEEP(double *net_radiation_d, double *soil_heat_flux_d, double *ndvi_d, double *surface_temperature_d, double *d0_d, double *kb1_d, double *zom_d, double *ustar_d, double *rah_d, double *H_d, float ndvi_max, float ndvi_min)
+__global__ void rah_correction_cycle_STEEP(double *net_radiation_d, double *soil_heat_flux_d, double *ndvi_d, double *surface_temperature_d, double *d0_d, double *kb1_d, double *zom_d, double *ustar_d, double *rah_d, double *H_d, double ndvi_max, double ndvi_min)
 {
     unsigned int pos = threadIdx.x + blockIdx.x * blockDim.x;
 
@@ -128,35 +128,35 @@ __global__ void rah_correction_cycle_STEEP(double *net_radiation_d, double *soil
         unsigned int hot_pos = hotEndmemberLine_d * width_d + hotEndmemberCol_d;
         unsigned int cold_pos = coldEndmemberLine_d * width_d + coldEndmemberCol_d;
 
-        float rah_ini_hot = rah_d[hot_pos];
-        float rah_ini_cold = rah_d[cold_pos];
+        double rah_ini_hot = rah_d[hot_pos];
+        double rah_ini_cold = rah_d[cold_pos];
 
-        float fc_hot = 1.0 - pow((ndvi_d[hot_pos] - ndvi_max) / (ndvi_min - ndvi_max), 0.4631);
-        float fc_cold = 1.0 - pow((ndvi_d[cold_pos] - ndvi_max) / (ndvi_min - ndvi_max), 0.4631);
+        double fc_hot = 1.0 - pow((ndvi_d[hot_pos] - ndvi_max) / (ndvi_min - ndvi_max), 0.4631);
+        double fc_cold = 1.0 - pow((ndvi_d[cold_pos] - ndvi_max) / (ndvi_min - ndvi_max), 0.4631);
 
-        float LE_hot = 0.55 * fc_hot * (net_radiation_d[hot_pos] - soil_heat_flux_d[hot_pos]) * 0.78;
-        float LE_cold = 1.75 * fc_cold * (net_radiation_d[cold_pos] - soil_heat_flux_d[cold_pos]) * 0.78;
+        double LE_hot = 0.55 * fc_hot * (net_radiation_d[hot_pos] - soil_heat_flux_d[hot_pos]) * 0.78;
+        double LE_cold = 1.75 * fc_cold * (net_radiation_d[cold_pos] - soil_heat_flux_d[cold_pos]) * 0.78;
 
-        float H_cold = net_radiation_d[cold_pos] - soil_heat_flux_d[cold_pos] - LE_hot;
-        float dt_cold = H_cold * rah_ini_cold / (RHO * SPECIFIC_HEAT_AIR);
+        double H_cold = net_radiation_d[cold_pos] - soil_heat_flux_d[cold_pos] - LE_hot;
+        double dt_cold = H_cold * rah_ini_cold / (RHO * SPECIFIC_HEAT_AIR);
 
-        float H_hot = net_radiation_d[hot_pos] - soil_heat_flux_d[hot_pos] - LE_cold;
-        float dt_hot = H_hot * rah_ini_hot / (RHO * SPECIFIC_HEAT_AIR);
+        double H_hot = net_radiation_d[hot_pos] - soil_heat_flux_d[hot_pos] - LE_cold;
+        double dt_hot = H_hot * rah_ini_hot / (RHO * SPECIFIC_HEAT_AIR);
 
-        float b = (dt_hot - dt_cold) / (surface_temperature_d[hot_pos] - surface_temperature_d[cold_pos]);
-        float a = dt_cold - (b * surface_temperature_d[cold_pos]);
+        double b = (dt_hot - dt_cold) / (surface_temperature_d[hot_pos] - surface_temperature_d[cold_pos]);
+        double a = dt_cold - (b * surface_temperature_d[cold_pos]);
 
         b_d = b;
         a_d = a;
 
-        float dt_final = a + b * surface_temperature_d[pos];
+        double dt_final = a + b * surface_temperature_d[pos];
         H_d[pos] = RHO * SPECIFIC_HEAT_AIR * dt_final / rah_d[pos];
-        float L = -1.0 * ((RHO * SPECIFIC_HEAT_AIR * pow(ustar_d[pos], 3.0) * surface_temperature_d[pos]) / (VON_KARMAN * GRAVITY * H_d[pos]));
+        double L = -1.0 * ((RHO * SPECIFIC_HEAT_AIR * pow(ustar_d[pos], 3.0) * surface_temperature_d[pos]) / (VON_KARMAN * GRAVITY * H_d[pos]));
 
-        float y2 = pow((1.0 - (16.0 * (10.0 - d0_d[pos])) / L), 0.25);
-        float x200 = pow((1.0 - (16.0 * (10.0 - d0_d[pos])) / L), 0.25);
+        double y2 = pow((1.0 - (16.0 * (10.0 - d0_d[pos])) / L), 0.25);
+        double x200 = pow((1.0 - (16.0 * (10.0 - d0_d[pos])) / L), 0.25);
 
-        float psi2, psi200;
+        double psi2, psi200;
         if (!isnan(L) && L > 0) {
             psi2 = -5.0 * ((10.0 - d0_d[pos]) / L);
             psi200 = -5.0 * ((10.0 - d0_d[pos]) / L);
@@ -167,14 +167,14 @@ __global__ void rah_correction_cycle_STEEP(double *net_radiation_d, double *soil
 
         ustar_d[pos] = (VON_KARMAN * ustar_d[pos]) / (logf((10.0 - d0_d[pos]) / zom_d[pos]) - psi200);
 
-        float rah_fst_part = 1.0 / (ustar_d[pos] * VON_KARMAN);
-        float rah_sec_part = logf((10.0 - d0_d[pos]) / zom_d[pos]) - psi2;
-        float rah_trd_part = rah_fst_part * kb1_d[pos];
+        double rah_fst_part = 1.0 / (ustar_d[pos] * VON_KARMAN);
+        double rah_sec_part = logf((10.0 - d0_d[pos]) / zom_d[pos]) - psi2;
+        double rah_trd_part = rah_fst_part * kb1_d[pos];
         rah_d[pos] = (rah_fst_part * rah_sec_part) + rah_trd_part;
     }
 }
 
-__global__ void rah_correction_cycle_ASEBAL(double *net_radiation_d, double *soil_heat_flux_d, double *ndvi_d, double *surface_temperature_d, double *kb1_d, double *zom_d, double *ustar_d, double *rah_d, double *H_d, float u200, int *stop_condition)
+__global__ void rah_correction_cycle_ASEBAL(double *net_radiation_d, double *soil_heat_flux_d, double *ndvi_d, double *surface_temperature_d, double *kb1_d, double *zom_d, double *ustar_d, double *rah_d, double *H_d, double u200, int *stop_condition)
 {
     unsigned int pos = threadIdx.x + blockIdx.x * blockDim.x;
 
@@ -182,31 +182,31 @@ __global__ void rah_correction_cycle_ASEBAL(double *net_radiation_d, double *soi
         unsigned int hot_pos = hotEndmemberLine_d * width_d + hotEndmemberCol_d;
         unsigned int cold_pos = coldEndmemberLine_d * width_d + coldEndmemberCol_d;
 
-        float rah_ini_hot = rah_d[hot_pos];
-        float rah_ini_cold = rah_d[cold_pos];
+        double rah_ini_hot = rah_d[hot_pos];
+        double rah_ini_cold = rah_d[cold_pos];
 
-        float H_cold = net_radiation_d[cold_pos] - soil_heat_flux_d[cold_pos];
-        float dt_cold = H_cold * rah_ini_cold / (RHO * SPECIFIC_HEAT_AIR);
+        double H_cold = net_radiation_d[cold_pos] - soil_heat_flux_d[cold_pos];
+        double dt_cold = H_cold * rah_ini_cold / (RHO * SPECIFIC_HEAT_AIR);
 
-        float H_hot = net_radiation_d[hot_pos] - soil_heat_flux_d[hot_pos];
-        float dt_hot = H_hot * rah_ini_hot / (RHO * SPECIFIC_HEAT_AIR);
+        double H_hot = net_radiation_d[hot_pos] - soil_heat_flux_d[hot_pos];
+        double dt_hot = H_hot * rah_ini_hot / (RHO * SPECIFIC_HEAT_AIR);
 
-        float b = (dt_hot - dt_cold) / (surface_temperature_d[hot_pos] - surface_temperature_d[cold_pos]);
-        float a = dt_cold - (b * surface_temperature_d[cold_pos]);
+        double b = (dt_hot - dt_cold) / (surface_temperature_d[hot_pos] - surface_temperature_d[cold_pos]);
+        double a = dt_cold - (b * surface_temperature_d[cold_pos]);
 
         b_d = b;
         a_d = a;
 
-        float dt_final = a + b * (surface_temperature_d[pos]);
+        double dt_final = a + b * (surface_temperature_d[pos]);
 
         H_d[pos] = RHO * SPECIFIC_HEAT_AIR * (dt_final) / rah_d[pos];
-        float L = -1.0 * ((RHO * SPECIFIC_HEAT_AIR * pow(ustar_d[pos], 3.0) * surface_temperature_d[pos]) / (VON_KARMAN * GRAVITY * H_d[pos]));
+        double L = -1.0 * ((RHO * SPECIFIC_HEAT_AIR * pow(ustar_d[pos], 3.0) * surface_temperature_d[pos]) / (VON_KARMAN * GRAVITY * H_d[pos]));
 
-        float x1 = pow((1.0 - (16.0 * 0.1) / L), 0.25);
-        float x2 = pow((1.0 - (16.0 * 2.0) / L), 0.25);
-        float x200 = pow((1.0 - (16.0 * 200.0) / L), 0.25);
+        double x1 = pow((1.0 - (16.0 * 0.1) / L), 0.25);
+        double x2 = pow((1.0 - (16.0 * 2.0) / L), 0.25);
+        double x200 = pow((1.0 - (16.0 * 200.0) / L), 0.25);
 
-        float psi1, psi2, psi200;
+        double psi1, psi2, psi200;
         if (!isnan(L) && L > 0) {
             psi1 = -5.0 * (0.1 / L);
             psi2 = -5.0 * (2.0 / L);
